Task error handling

+++ src/main/java/com/softserve/itacademy/service/impl/TaskServiceImpl.java
@@ -1,43 +1,75 @@
 package com.softserve.itacademy.service.impl;
 
+import com.softserve.itacademy.exception.NullEntityReferenceException;
 import com.softserve.itacademy.model.Task;
+import com.softserve.itacademy.model.ToDo;
 import com.softserve.itacademy.repository.TaskRepository;
+import com.softserve.itacademy.repository.ToDoRepository;
 import com.softserve.itacademy.service.TaskService;
+import org.springframework.dao.InvalidDataAccessApiUsageException;
 import org.springframework.stereotype.Service;
 
+import javax.persistence.EntityNotFoundException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
 @Service
 public class TaskServiceImpl implements TaskService {
-    private TaskRepository taskRepository;
+    private final TaskRepository taskRepository;
+    private final ToDoRepository todoRepository;
 
-    public TaskServiceImpl(TaskRepository taskRepository) {
+    public TaskServiceImpl(TaskRepository taskRepository, ToDoRepository todoRepository) {
         this.taskRepository = taskRepository;
+        this.todoRepository = todoRepository;
     }
 
     @Override
     public Task create(Task user) {
+        try {
             return taskRepository.save(user);
+        } catch (InvalidDataAccessApiUsageException | IllegalArgumentException e) {
+            throw new NullEntityReferenceException("Task cannot be null");
+        }
     }
 
     @Override
     public Task readById(long id) {
         Optional<Task> optional = taskRepository.findById(id);
+        if (optional.isPresent()) {
             return optional.get();
+        }
+        throw new EntityNotFoundException("Task with id " + id + " not found");
     }
 
     @Override
     public Task update(Task task) {
-            Task oldTask = readById(task.getId());
-                return taskRepository.save(task);
+        if (task != null) {
+            Task oldTask;
+            try {
+                oldTask = readById(task.getId());
+            } catch (IllegalArgumentException e) {
+                throw new NullEntityReferenceException("Task id cannot be null");
+            }
+            if (oldTask != null) {
+                try {
+                    return taskRepository.save(task);
+                } catch (InvalidDataAccessApiUsageException | IllegalArgumentException e) {
+                    throw new NullEntityReferenceException("Task cannot be null");
+                }
+            }
+        }
+        throw new NullEntityReferenceException("Task can`t be 'null'");
     }
 
     @Override
     public void delete(long id) {
         Task task = readById(id);
+        if (task != null) {
             taskRepository.delete(task);
+        } else {
+            throw new EntityNotFoundException("Task with id " + id + " not found");
+        }
     }
 
     @Override
@@ -48,7 +80,12 @@ public class TaskServiceImpl implements TaskService {
 
     @Override
     public List<Task> getByTodoId(long todoId) {
-        List<Task> tasks = taskRepository.getByTodoId(todoId);
-        return tasks.isEmpty() ? new ArrayList<>() : tasks;
+        Optional<ToDo> optional = todoRepository.findById(todoId);
+        if (optional.isPresent()) {
+            List<Task> tasks = taskRepository.getByTodoId(todoId);
+            return tasks.isEmpty() ? new ArrayList<>() : tasks;
+        } else {
+            throw new EntityNotFoundException("Todo with id " + todoId + " not found");
+        }
     }
 }
